Internal cleanup, attempt to test deep_explain
No longer declare Base* types, which could leak out, and fold
ResultSource into the dynamic type generation loop to reduce code
duplication.

Added a test for deep_explain, which apparently is only called if the
type  must be parameterized- if it is optionally not parameterized, it
isn't called. So the test is entirely skipped. Revisit if and when
https://rt.cpan.org/Public/Bug/Display.html?id=104154 is fixed (or I
figure out "the right way")